Use custom data structures to (de)serialize RPC to acquire the best performance #322
Former-commit-id: f446e8dad471db63a4d94224fa85aebcc1e27863

+++ turms/src/main/java/im/turms/turms/access/web/controller/user/UserOnlineInfoController.java
@@ -110,7 +110,7 @@ public class UserOnlineInfoController {
     public Mono<ResponseEntity<ResponseDTO<Collection<User>>>> queryUsersNearby(
             @RequestParam Long userId,
             @RequestParam(required = false) DeviceType deviceType,
-            @RequestParam(required = false) Integer maxPeopleNumber,
+            @RequestParam(required = false) Short maxPeopleNumber,
             @RequestParam(required = false) Double maxDistance) {
         Flux<User> usersNearby = usersNearbyService.queryUsersProfilesNearby(userId, deviceType, maxPeopleNumber, maxDistance);
         return ResponseFactory.okIfTruthy(usersNearby);
